Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pytz==2023.3.post1
Werkzeug==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
python-dotenv==1.0.0
mysqlclient==2.2.0
FLask-Migrate==4.0.5